Filter (some) false positives
When potential hazards are detected, the hazard detector will
write "hazard detected", which can be used to stop the rover.
The traversability map however will not be written yet, which
means that the path will not be replanned immediately.
Instead, we now, while the rover is stopped, count how often each
pixel is considered a hazard and ignore those pixels, which are
spurious.

The number of frames we take into consideration and the threshold
should become configuration parameters.