Add WIP logic to allow Ackermann maneuvers
Once an obstacle is detected, we wait for a new plan from
the path planner.
If this plan is a point turn, we'll allow it as before,
but now we also allow performing commands with positive translation
to be executed for a certain (TBD) amount of time, while ignoring
the part of the hazard detector's region of interest containing the
obstacle.
Once this time has run out, we will inspect the whole RoI for potential
hazards again.