Move interpreting new plans to a different state
Before this commit, the motion command of a new plan would
be read when transitioning from WAITING_FOR_NEW_PLAN to
AVOIDANCE_MANEUVER as soon as the path planner signals that
a new plan is available.

A problem with this is that the signal new_plan is available
while the command is still the previous STOP, so we cannot
infer which side of the RoI to ignore and in the next loop
the new_plan signal is not "NEW" anymore.

Now the function trustAvoidanceManeuver is moved to the state
AVOIDANCE_MANEUVER and waits for the first non-stop command,
resetting the wait timer until then.

Co-Authored-By: Martin Azkarate <martin.azkarate@gmail.com>